Move backend main loop into new serve module
